+++ crates/testing-utils/src/tss_server_process.rs
@@ -13,15 +13,10 @@
 // You should have received a copy of the GNU Affero General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use std::{net::TcpListener, time::Duration};
-
 use axum::{routing::IntoMakeService, Router};
 use entropy_kvdb::{encrypted_sled::PasswordMethod, kv_manager::KvManager};
 use entropy_protocol::{KeyParams, PartyId};
-use rand_core::OsRng;
-use subxt::utils::AccountId32 as SubxtAccountId32;
-use synedrion::KeyShare;
-
+use entropy_shared::DETERMINISTIC_KEY_SHARE;
 use entropy_tss::{
     app,
     get_signer,
@@ -29,6 +24,10 @@ use entropy_tss::{
     // signing_client::ListenerState,
     AppState,
 };
+use rand_core::OsRng;
+use std::{net::TcpListener, time::Duration};
+use subxt::utils::AccountId32 as SubxtAccountId32;
+use synedrion::{ecdsa::SigningKey, KeyShare};
 
 pub const DEFAULT_ENDPOINT: &str = "ws://localhost:9944";
 
@@ -60,9 +59,11 @@ async fn create_clients(
 }
 
 pub async fn spawn_testing_validators(
-    sig_req_keyring: Option<String>,
+    passed_verifying_key: Option<Vec<u8>>,
     // If this is true a keyshare for the user will be generated and returned
     extra_private_keys: bool,
+    // If true keyshare and verifying key is deterministic
+    deterministic_key_share: bool,
 ) -> (Vec<String>, Vec<PartyId>, Option<KeyShare<KeyParams>>) {
     // spawn threshold servers
     let ports = [3001i64, 3002];
@@ -79,20 +80,37 @@ pub async fn spawn_testing_validators(
         *get_signer(&bob_kv).await.unwrap().account_id().clone().as_ref(),
     ));
 
-    let user_keyshare_option = if sig_req_keyring.is_some() {
+    let user_keyshare_option = if passed_verifying_key.is_some() {
         let number_of_shares = if extra_private_keys { 3 } else { 2 };
-        let shares = KeyShare::<KeyParams>::new_centralized(&mut OsRng, number_of_shares, None);
+        // creates a deterministic keyshare if requiered
+        let signing_key = if deterministic_key_share {
+            Some(SigningKey::from_bytes((&*DETERMINISTIC_KEY_SHARE).into()).unwrap())
+        } else {
+            None
+        };
+
+        let shares = KeyShare::<KeyParams>::new_centralized(
+            &mut OsRng,
+            number_of_shares,
+            signing_key.as_ref(),
+        );
         let validator_1_threshold_keyshare: Vec<u8> =
             entropy_kvdb::kv_manager::helpers::serialize(&shares[0]).unwrap();
         let validator_2_threshold_keyshare: Vec<u8> =
             entropy_kvdb::kv_manager::helpers::serialize(&shares[1]).unwrap();
+
+        // uses the deterministic verifying key if requested
+        let verifying_key = if deterministic_key_share {
+            hex::encode(shares[0].verifying_key().to_encoded_point(true).as_bytes())
+        } else {
+            hex::encode(passed_verifying_key.unwrap())
+        };
+
         // add key share to kvdbs
-        let alice_reservation =
-            alice_kv.kv().reserve_key(sig_req_keyring.clone().unwrap()).await.unwrap();
+        let alice_reservation = alice_kv.kv().reserve_key(verifying_key.clone()).await.unwrap();
         alice_kv.kv().put(alice_reservation, validator_1_threshold_keyshare).await.unwrap();
 
-        let bob_reservation =
-            bob_kv.kv().reserve_key(sig_req_keyring.clone().unwrap()).await.unwrap();
+        let bob_reservation = bob_kv.kv().reserve_key(verifying_key.clone()).await.unwrap();
         bob_kv.kv().put(bob_reservation, validator_2_threshold_keyshare).await.unwrap();
 
         if extra_private_keys {