+++ CHANGELOG.md
@@ -35,6 +35,9 @@ At the moment this project **does not** adhere to
   the Registry pallet as this better describes the purpose of the pallet.
 - In [#662](https://github.com/entropyxyz/entropy-core/pull/662), the Free Transaction pallet was
   removed.
+- In [#666](https://github.com/entropyxyz/entropy-core/pull/666), Permissioned access type was removed 
+  to be handled by a program. Also in `registered` `sig_request_key` as the key for the struct was 
+  replaced by `verifying_key`. This means `verifying_key` was removed from the `registered` struct.
 
 ### Added
 - Add ValidatorSubgroupRotated event ([#618](https://github.com/entropyxyz/entropy-core/pull/618))
@@ -50,6 +53,7 @@ At the moment this project **does not** adhere to
 - Disallow using existing TSS account IDs in Staking pallet ([#657](https://github.com/entropyxyz/entropy-core/pull/657))
 - Clean ups around Staking Extension's `validate()` extrinsic ([#659](https://github.com/entropyxyz/entropy-core/pull/659))
 - Rename `pallet_relayer` to `pallet_registry` ([#661](https://github.com/entropyxyz/entropy-core/pull/661))
+- Remove permissioned access type ([#666](https://github.com/entropyxyz/entropy-core/pull/666))
 - Use SessionID in shared randomness ([#676](https://github.com/entropyxyz/entropy-core/pull/676))
 
 ### Removed